게시판 만들기 코드 보안 컴포넌트 분리 및 설명하는 주석 추가
