+++ src/noticeBoard/NoticeBoard.tsx
@@ -1,14 +1,16 @@
 import React, {useEffect, useState} from 'react';
 import './NoticeBoard.css'
+import {Post} from "../type/Post";
+import PostList from "./PostList";
+import Pagination from "./Pagination";
 
-interface Post {
-	id:number,
-	title:string
-	content:string,
-}
 
 const _postPerPage: number = 5;//1페이지에 게시물 몇개?
 
+/**
+ * 게시판 전체
+ * @constructor
+ */
 export default function NoticeBoard() {
 	const [postList, setPostList] = useState<Post[]>([]) //게시글 데이터
 	const [currentPage, setCurrentPage] = useState<number>(1);
@@ -42,68 +44,5 @@ export default function NoticeBoard() {
 	);
 }
 
-interface PostListProps {
-	posts: Post[]
-}
-
-function PostList({posts}: PostListProps) {
-	return (
-		<div className="post-list">
-			{posts.map((post) => {
-				return(
-					<div className="post" key={'post'+post.id}>
-						<h2>{post.title}</h2>
-						<p>{post.content}</p>
-					</div>
-				)
-			})}
-		</div>
-	)
-}
 
 
-interface PaginationProps {
-	paginate:(pageNumber:number) => void
-	totalPosts:number
-	postPerPage:number
-	currentPage:number
-}
-
-
-function Pagination({paginate,totalPosts,postPerPage,currentPage}:PaginationProps){
-	const totalPages = Math.ceil(totalPosts/postPerPage)//페이지 갯수
-	const [pageGroup, setPageGroup] = useState(1);//표시할 페이지 그룹
-
-	//전체 페이지 얕은 복사
-	const pages = Array.from({length: totalPages}, (_, i) => i+1);
-
-	const startPage = (pageGroup - 1) * 5; //페이지 앞
-	const endPage = startPage + 4; //페이지 끝
-	const visiblePages = pages.slice(startPage, endPage + 1);//보여지는 페이지
-
-	const handlePageClick = (page:number):void => {
-		if(page<=0 || page>totalPages) {
-			return;
-		}
-		paginate(page)
-		const newPageGroup = Math.ceil(page/5);
-		setPageGroup(newPageGroup);
-	}
-
-
-	return(
-		<div className="pagination">
-			{pageGroup ===1 ||
-				(<button
-					onClick={() => handlePageClick((pageGroup-1)*5)}>{'<'}</button>)}
-			{visiblePages.map((page) => {
-				return(
-					<button style={currentPage===page?{backgroundColor:'#2222'}:{}} key={page} onClick={()=>paginate(page)}>{page}</button>
-				)
-			})}
-			{pageGroup === Math.ceil(totalPages/5) || (<button
-				onClick={() => handlePageClick(pageGroup*5+1)}>{'>'}</button>)}
-
-		</div>
-	)
-}
