+++ src/type/Post.ts
@@ -1,0 +1,6 @@
+export interface Post {
+	id:number,
+	title:string
+	content:string,
+}
+
